+++ libs/components/src/lib/data-grid/data-grid-cell.ts
@@ -1,4 +1,4 @@
-import { DataGridCell as FoundationElement } from '@microsoft/fast-foundation';
+import { DataGridCell as FoundationDataGridCell } from '@microsoft/fast-foundation';
 import {attr} from '@microsoft/fast-element';
 
 /**
@@ -6,12 +6,17 @@ import {attr} from '@microsoft/fast-element';
  *
  * @public
  */
-export class DataGridCell extends FoundationElement {
+export class DataGridCell extends FoundationDataGridCell {
 
 	@attr({attribute: 'aria-selected', mode: 'fromView'}) override ariaSelected: string | null = null;
 
 	ariaSelectedChanged(_: string | null, selectedState: string | null) {
 		this.shadowRoot!.querySelector('.base')?.classList.toggle('selected', selectedState === 'true');
+	}
+
+	override connectedCallback() {
+		super.connectedCallback();
+		this.ariaSelectedChanged(null, this.ariaSelected);
 	}
 
 	override handleFocusin(e: FocusEvent) {
@@ -22,5 +27,16 @@ export class DataGridCell extends FoundationElement {
 	override handleFocusout(e: FocusEvent) {
 		super.handleFocusout(e);
 		this.shadowRoot!.querySelector('.base')!.classList.remove('active');
+	}
+
+	constructor() {
+		super();
+		(this as any).updateCellStyle = () => {
+			if (this.gridColumn && !this.gridColumn.includes('undefined')){
+				this.style.gridColumn = this.gridColumn;
+			} else {
+				this.style.removeProperty('grid-column');
+			}
+		};
 	}
 }