+++ libs/components/src/lib/data-grid/data-grid-row.ts
@@ -1,9 +1,12 @@
-import { DataGridRow as FoundationElement } from '@microsoft/fast-foundation';
+import { DataGridRow as FoundationDataGridRow } from '@microsoft/fast-foundation';
+import {attr} from '@microsoft/fast-element';
 
 /**
  * Base class for data-grid
  *
  * @public
  */
-export class DataGridRow extends FoundationElement {
+export class DataGridRow extends FoundationDataGridRow {
+	@attr({attribute: 'aria-selected'})
+	override ariaSelected: string | null = null;
 }