+++ libs/components/src/lib/data-grid/index.spec.ts
@@ -10,6 +10,9 @@ describe('data grid integration tests', () => {
 		return element.rowElements[row].children[cell] as HTMLElement;
 	}
 
+	function getRow(row: number) {
+		return element.rowElements[row] as HTMLElement;
+	}
 	let element: DataGrid;
 
 	beforeEach(async () => {
@@ -77,6 +80,24 @@ describe('data grid integration tests', () => {
 
 	describe('selectionMode', function () {
 
+		it('should remove aria-selected when "none" is set initially', async function () {
+			element = (await fixture(
+				`<${COMPONENT_TAG} selection-mode="none">
+											<vwc-data-grid-row aria-selected="true">
+												<vwc-data-grid-cell aria-selected="true">Cell 1</vwc-data-grid-cell>
+												<vwc-data-grid-cell aria-selected="false">Cell 2</vwc-data-grid-cell>
+											</vwc-data-grid-row>
+											<vwc-data-grid-row aria-selected="false">
+												<vwc-data-grid-cell aria-selected="true">Cell 1</vwc-data-grid-cell>
+												<vwc-data-grid-cell aria-selected="false">Cell 2</vwc-data-grid-cell>
+											</vwc-data-grid-row>
+										</${COMPONENT_TAG}>`
+			)) as DataGrid;
+			await elementUpdated(element);
+			const ariaSelectedExists = element.querySelectorAll('[aria-selected]').length > 0;
+			expect(ariaSelectedExists).toEqual(false);
+		});
+
 		describe('cell-selection', function () {
 			let cell: HTMLElement;
 
@@ -89,6 +110,7 @@ describe('data grid integration tests', () => {
 				await elementUpdated(element);
 				cell = getRowCell(1,0);
 			});
+
 			it.each([DataGridSelectionMode.singleCell, DataGridSelectionMode.multiCell])
 			('should add aria-selected="false" to all cells if selectionMode is %s', async function (selectionMode: DataGridSelectionMode) {
 				element.selectionMode = selectionMode;
@@ -105,9 +127,21 @@ describe('data grid integration tests', () => {
 				element.selectionMode = DataGridSelectionMode.none;
 				await elementUpdated(element);
 
-				const allNoneHeaderCells = Array.from(element.querySelectorAll('[role="gridcell"]'));
-				const allNoneHeaderCellsNotSelectable = allNoneHeaderCells.every(cell => !cell.hasAttribute('aria-selected'));
-				expect(allNoneHeaderCellsNotSelectable).toEqual(true);
+				const allNonHeaderCells = Array.from(element.querySelectorAll('[role="gridcell"]'));
+				const allNonHeaderCellsNotSelectable = allNonHeaderCells.every(cell => !cell.hasAttribute('aria-selected'));
+				expect(allNonHeaderCellsNotSelectable).toEqual(true);
+			});
+
+			it('should remove aria-selected from all rows if set to "*-cell"', async function () {
+				element.selectionMode = DataGridSelectionMode.singleRow;
+				await elementUpdated(element);
+
+				element.selectionMode = DataGridSelectionMode.singleCell;
+				await elementUpdated(element);
+
+				const allNonHeaderRows = Array.from(element.querySelectorAll('[role="row"]'));
+				const allNonHeaderRowsNotSelectable = allNonHeaderRows.every(row => !row.hasAttribute('aria-selected'));
+				expect(allNonHeaderRowsNotSelectable).toEqual(true);
 			});
 
 			it.each([DataGridSelectionMode.singleCell, DataGridSelectionMode.multiCell])
@@ -220,6 +254,122 @@ describe('data grid integration tests', () => {
 				expect(isElementSelected(cell1)).toEqual(true);
 				expect(isElementSelected(cell2)).toEqual(true);
 			});
+
+			it('should leave aria-selected value to cells with aria-selected on init', async function () {
+				element = (await fixture(
+					`<${COMPONENT_TAG} selection-mode="single-cell">
+											<vwc-data-grid-row>
+												<vwc-data-grid-cell aria-selected="true">Cell 1</vwc-data-grid-cell>
+												<vwc-data-grid-cell>Cell 2</vwc-data-grid-cell>
+											</vwc-data-grid-row>
+										</${COMPONENT_TAG}>`
+				)) as DataGrid;
+				await elementUpdated(element);
+				const cell = element.querySelector('vwc-data-grid-cell') as HTMLElement;
+				expect(cell.getAttribute('aria-selected')).toEqual('true');
+			});
+		});
+
+		describe('row selection', () => {
+			beforeEach(async function () {
+				element.rowsData = [
+					{ id: 1, name: 'John', age: 20 },
+					{ id: 2, name: 'Jane', age: 21 },
+				];
+				await elementUpdated(element);
+				await elementUpdated(element);
+			});
+
+			it('should remove aria-selected from all rows if set to "none"', async function () {
+				element.selectionMode = DataGridSelectionMode.multiRow;
+				await elementUpdated(element);
+
+				element.selectionMode = DataGridSelectionMode.none;
+				await elementUpdated(element);
+
+				const allNoneHeaderRows = Array.from(element.querySelectorAll('[role="row"]'));
+				const allNonHeaderRowsNotSelectable = allNoneHeaderRows.every(row => !row.hasAttribute('aria-selected'));
+				expect(allNonHeaderRowsNotSelectable).toEqual(true);
+			});
+
+			it('should remove aria-selected from all cells if set to "*-row"', async function () {
+				element.selectionMode = DataGridSelectionMode.singleCell;
+				await elementUpdated(element);
+
+				element.selectionMode = DataGridSelectionMode.multiRow;
+				await elementUpdated(element);
+
+				const allNonHeaderCells = Array.from(element.querySelectorAll('[role="gridcell"]'));
+				const allNonHeaderCellsNotSelectable = allNonHeaderCells.every(cell => !cell.hasAttribute('aria-selected'));
+				expect(allNonHeaderCellsNotSelectable).toEqual(true);
+			});
+
+			it.each([DataGridSelectionMode.singleRow, DataGridSelectionMode.multiRow])
+			('should set aria-selected="false" on all non header rows if selectionMode is %s',
+				async function (selectionMode: DataGridSelectionMode) {
+					element.selectionMode = selectionMode;
+					await elementUpdated(element);
+					const allNonHeaderRows = element.querySelectorAll('[role="row"]');
+					const allNonHeaderRowsHaveSelectedFalse = Array.from(allNonHeaderRows).every((row) => {
+						return row.getAttribute('aria-selected') === 'false';
+					});
+					expect(allNonHeaderRowsHaveSelectedFalse).toEqual(true);
+				});
+
+			it('should set aria-selected="true" on clicked row', async function () {
+				element.selectionMode = DataGridSelectionMode.singleRow;
+				const cell = getRowCell(1, 0);
+				const row = getRow(1);
+				cell.click();
+				await elementUpdated(element);
+				expect(row.getAttribute('aria-selected')).toEqual('true');
+			});
+
+			it('should set aria-selected="true" on clicked row and remove from other rows', async function () {
+				element.selectionMode = DataGridSelectionMode.singleRow;
+				const row1 = getRow(1);
+				const cellInRow1 = getRowCell(1, 0);
+				const row2 = getRow(2);
+				const cellInRow2 = getRowCell(2, 0);
+
+				cellInRow1.click();
+				await elementUpdated(element);
+				cellInRow2.click();
+				await elementUpdated(element);
+
+				expect(row1.getAttribute('aria-selected')).toEqual('false');
+				expect(row2.getAttribute('aria-selected')).toEqual('true');
+			});
+
+			it.each(['ctrlKey', 'shiftKey', 'metaKey'])
+			('should set aria-selected="true" to all clicked rows in "multi-row" state and %s key pressed',
+				function (activeKey) {
+					element.selectionMode = DataGridSelectionMode.multiRow;
+					const row1 = getRow(1);
+					const cellInRow1 = getRowCell(1, 0);
+					const row2 = getRow(2);
+					const cellInRow2 = getRowCell(2, 0);
+
+					cellInRow1.click();
+					cellInRow2.dispatchEvent(new MouseEvent('click', { [activeKey]: true, bubbles: true, composed: true }));
+
+					expect(row1.getAttribute('aria-selected')).toEqual('true');
+					expect(row2.getAttribute('aria-selected')).toEqual('true');
+				});
+
+			it('should leave aria-selected value to rows on init', async function () {
+				element = (await fixture(
+					`<${COMPONENT_TAG} selection-mode="single-row">
+											<vwc-data-grid-row aria-selected="true">
+												<vwc-data-grid-cell>Cell 1</vwc-data-grid-cell>
+												<vwc-data-grid-cell>Cell 2</vwc-data-grid-cell>
+											</vwc-data-grid-row>
+										</${COMPONENT_TAG}>`
+				)) as DataGrid;
+				await elementUpdated(element);
+				const row = element.querySelector('vwc-data-grid-row') as HTMLElement;
+				expect(row.getAttribute('aria-selected')).toEqual('true');
+			});
 		});
 	});
 
@@ -229,14 +379,9 @@ function isElementSelected(element: HTMLElement): boolean {
 	return element.getAttribute('aria-selected') === 'true';
 }
 
+// TODO:: add "manual" example + remove "grid-column" if `undefined`
 // TODO:: header cells design - sort and filter variants
-// TODO:: row selection
-// TODO:: aria-selected to selected rows
 // TODO:: header cell with filter
 // TODO:: header cell with sort
 // TODO:: add sorting and filtering examples
-// TODO:: add "manual" example
 // TODO:: test keyboard navigation
-// TODO:: add aria-multiselectable to grid if multi selection is enabled
-// TODO:: add aria-selected to grid rows only if selection is enabled (remove when changing selection mode)
-// TODO:: check the rtl states