+++ libs/components/src/lib/select/select.template.ts
@@ -16,14 +16,15 @@ const getStateClasses = ({
 	appearance,
 	metaSlottedContent,
 	errorValidationMessage,
-	successText
+	successText,
 }: Select) => classNames(
 	['disabled', disabled],
 	[`appearance-${appearance}`, Boolean(appearance)],
 	[`shape-${shape}`, Boolean(shape)],
 	['has-meta', Boolean(metaSlottedContent?.length)],
 	['error connotation-alert', Boolean(errorValidationMessage)],
-	['success connotation-success', !!successText]
+	['success connotation-success', !!successText],
+	['has-meta', Boolean(metaSlottedContent?.length)],
 );
 
 /**
@@ -37,7 +38,7 @@ function renderLabel() {
 }
 
 function selectValue(context: ElementDefinitionContext) {
-	const affixIconTemplate = affixIconTemplateFactory(context);
+	const affixIconTemplate = affixIconTemplateFactory(context, false);
 	const focusTemplate = focusTemplateFactory(context);
 	return html<Select>`
 		<div
@@ -47,7 +48,9 @@ function selectValue(context: ElementDefinitionContext) {
 					${ref('_anchor')}
 				>
 					<div class="selected-value">
-						${x => affixIconTemplate(x.icon)}
+						<slot name="icon">
+							${when(x => x.icon, html<Select>`${x => affixIconTemplate(x.icon)}`)}
+						</slot>
 						<span class="text">${x => x.displayValue}</span>
 						<slot name="meta" ${slotted('metaSlottedContent')}></slot>
 					</div>
@@ -55,6 +58,10 @@ function selectValue(context: ElementDefinitionContext) {
 					${() => focusTemplate}
 				</div>
 	`;
+}
+
+function setFixedDropdownVarWidth(x: Select) {
+	return (x.open && x.fixedDropdown) ? `--_select-fixed-width: ${Math.round(x.getBoundingClientRect().width)}px` : null;
 }
 
 /**
@@ -72,9 +79,10 @@ function renderControl(context: ElementDefinitionContext) {
 					?open="${x => (x.collapsible ? x.open : true)}"
 					anchor="control"
 					placement="bottom-start"
-							strategy="absolute"
+							strategy="${x => x.fixedDropdown ? null : 'absolute'}"
 							${ref('_popup')}
 							class="popup"
+					style="${setFixedDropdownVarWidth}"
 							>
 							<div
                 id="${x => x.listboxId}"
@@ -116,23 +124,25 @@ export const SelectTemplate: (
 
 	return html<Select>`
 	  <template class="base"
-            aria-activedescendant="${x => x.ariaActiveDescendant}"
-            aria-controls="${x => x.ariaControls}"
-            aria-disabled="${x => x.ariaDisabled}"
-            aria-expanded="${x => x.ariaExpanded}"
-            aria-haspopup="${x => (x.collapsible ? 'listbox' : null)}"
-            aria-multiselectable="${x => x.ariaMultiSelectable}"
-            ?open="${x => x.open}"
-            role="select"
-            tabindex="${x => (!x.disabled ? '0' : null)}"
-            @click="${(x, c) => x.clickHandler(c.event as MouseEvent)}"
-            @focusin="${(x, c) => x.focusinHandler(c.event as FocusEvent)}"
-            @focusout="${(x, c) => x.focusoutHandler(c.event as FocusEvent)}"
-            @keydown="${(x, c) => x.keydownHandler(c.event as KeyboardEvent)}"
-            @mousedown="${(x, c) => x.mousedownHandler(c.event as MouseEvent)}"
+							aria-label="${x => x.ariaLabel ? x.ariaLabel : x.label}"
+							aria-activedescendant="${x => x.ariaActiveDescendant}"
+							aria-controls="${x => x.ariaControls}"
+							aria-disabled="${x => x.ariaDisabled}"
+							aria-expanded="${x => x.ariaExpanded}"
+							aria-haspopup="${x => (x.collapsible ? 'listbox' : null)}"
+							aria-multiselectable="${x => x.ariaMultiSelectable}"
+							?open="${x => x.open}"
+							role="combobox"
+							tabindex="${x => (!x.disabled ? '0' : null)}"
+							@click="${(x, c) => x.clickHandler(c.event as MouseEvent)}"
+							@focusin="${(x, c) => x.focusinHandler(c.event as FocusEvent)}"
+							@focusout="${(x, c) => x.focusoutHandler(c.event as FocusEvent)}"
+							@keydown="${(x, c) => x.keydownHandler(c.event as KeyboardEvent)}"
+							@mousedown="${(x, c) => x.mousedownHandler(c.event as MouseEvent)}"
         >
             ${renderControl(context)}
 		</template>
-// TODO:: add focus to the control-wrapper
 	`;
 };
+
+// TODO::change the css variable according to select width