+++ libs/components/src/lib/tabs/tabs.spec.ts
@@ -1,19 +1,27 @@
-import { elementUpdated, fixture, getBaseElement } from '@vivid-nx/shared';
-import type { Tab } from '../tab/tab';
-import { Tabs } from './tabs';
+import {elementUpdated, fixture, getBaseElement} from '@vivid-nx/shared';
+import {Connotation} from '@vonage/vivid';
+import type {Tab} from '../tab/tab';
+import {Tabs} from './tabs';
 import '../tab-panel/tab-panel';
 import '.';
 
 const COMPONENT_TAG = 'vwc-tabs';
 
-window.HTMLElement.prototype.getBoundingClientRect = function () {
-	return {
-		x: 146, y: 50, width: 440, height: 240,
-		top: 50, right: 586, bottom: 290, left: 146
-	} as DOMRect;
-};
-
 describe('vwc-tabs', () => {
+
+	beforeEach(function () {
+		window.HTMLElement.prototype.getBoundingClientRect = function () {
+			return {
+				x: 146, y: 50, width: 440, height: 240,
+				top: 50, right: 586, bottom: 290, left: 146
+			} as DOMRect;
+		};
+	});
+
+	afterEach(function () {
+		window.HTMLElement.prototype.getBoundingClientRect = jest.fn();
+	});
+
 	async function setFixture(activeid: string | null = 'apps'): Promise<Tabs> {
 		return (await fixture(`<${COMPONENT_TAG} ${activeid ? `activeid="${activeid}"` : ''}>
 		<vwc-tab label="Appetizers" id="apps"></vwc-tab>
@@ -71,7 +79,6 @@ describe('vwc-tabs', () => {
 		});
 	});
 
-
 	describe('activeid', () => {
 		it('should set activeid property', async () => {
 			const activeid = 'entrees';
@@ -89,6 +96,60 @@ describe('vwc-tabs', () => {
 			const tmpElement = await setFixture(null);
 			await elementUpdated(tmpElement);
 			expect(tmpElement.activeid).toEqual('apps');
+		});
+	});
+
+	describe('connotation', () => {
+		function checkConnotationOnActiveTab(connotation: Connotation = Connotation.CTA) {
+			expect(element.activetab.getAttribute('connotation')).toEqual(connotation);
+		}
+
+		function checkConnotationDoesntExistOnNonActiveTabs() {
+			const nonActiveTabs = Array.from(element.querySelectorAll('vwc-tab:not([aria-selected="true"])'));
+			nonActiveTabs.forEach(tab => {
+				expect(tab.hasAttribute('connotation')).toBeFalsy();
+			});
+		}
+		beforeEach(function () {
+			element.connotation = Connotation.CTA;
+		});
+
+		it('should reflect connotation on active tab after init', async function () {
+			await elementUpdated(element);
+			checkConnotationOnActiveTab();
+			checkConnotationDoesntExistOnNonActiveTabs();
+		});
+
+		it('should reflect connotation on active tab after activeid changed', async function () {
+			element.activeid = 'entrees';
+			await elementUpdated(element);
+			checkConnotationOnActiveTab();
+			checkConnotationDoesntExistOnNonActiveTabs();
+		});
+
+		it('should reflect connotation on active tab after orientation changed', async function () {
+			element.orientation = 'vertical';
+			await elementUpdated(element);
+			checkConnotationOnActiveTab();
+			checkConnotationDoesntExistOnNonActiveTabs();
+		});
+
+		it('should reflect connotation on active tab after tabs changed', async function () {
+			const newTab = document.createElement('vwc-tab');
+			newTab.slot = 'tab';
+			element.appendChild(newTab);
+			await elementUpdated(element);
+			checkConnotationOnActiveTab();
+			checkConnotationDoesntExistOnNonActiveTabs();
+		});
+
+		it('should reflect connotation on active tab after tab panels changed', async function () {
+			const newTabPanel = document.createElement('vwc-tab-panel');
+			newTabPanel.slot = 'tabpanel';
+			element.appendChild(newTabPanel);
+			await elementUpdated(element);
+			checkConnotationOnActiveTab();
+			checkConnotationDoesntExistOnNonActiveTabs();
 		});
 	});